+++ CHANGELOG.md
@@ -2,14 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
-# [UnReleased]
+# [Released]
 
 ## [v1.1.5.6]
 
 ### Added
     - File Manager now displays the success/failure/last stats from OctoPrint
-    - New API Endpoint for collecting enviromental data. Expected the following JSON format: {temperature: value (°C), pressure: value (hPa), humidity:value (%), gas_resistance: value (KOhms), date: value}. Null values are required when no data present. 
- 
+    - New API Endpoint for collecting enviromental data. Expected the following JSON format: {temperature: value (°C / °F), pressure: value (hPa), humidity:value (%), gas_resistance: value (KOhms), date: value}. Null values are required when no data present. 
+    - Dashboard will now show envriomental history... currently shows hard coded last 1500 records.
  
 ### Changed